requirements.txt
# 基本依赖
pillow>=10.0.0
#customtkinter>=5.2.0
pyinstaller==6.12.0  # 使用实际存在的最新版本
pywin32; sys_platform == 'win32'  # Windows 特定依赖

# 加密和安全依赖
pyOpenSSL>=23.0.0
cryptography>=39.0.0

protobuf>=3.19.0
netifaces>=0.11.0

# 可能需要的其他依赖
pefile; sys_platform == 'win32'  # Windows PE文件支持
setuptools  # 一些 PyInstaller 功能可能需要


